backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
pandas==2.1.3
numpy==1.26.2
apscheduler==3.10.4
loguru==0.7.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
websockets==12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psutil==5.9.6
aiosmtplib==1.2.0
email-validator==2.1.0.post1
